requirements.txt
Django
djangorestframework
django-phonenumber-field
django-phonenumbers
djangorestframework
drf-nested-routers
model_bakery
pytest-django
djangorestframework_simplejwt